refactor(*): place entites inside domain directory
