+++ src/main/java/com/battlecruisers/yanullja/room/domain/Room.java
@@ -1,7 +1,6 @@
-package com.battlecruisers.yanullja.room;
+package com.battlecruisers.yanullja.room.domain;
 
-import com.battlecruisers.yanullja.place.Place;
-import com.battlecruisers.yanullja.subregion.SubRegion;
+import com.battlecruisers.yanullja.place.domain.Place;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
